+++ src/bin/mnp.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-import { _version, _help, _init, _name, _check, _delete, _scope, _noScope, _struct, _desc } from './get-args'
+import { _version, _help, _init, _name, _check, _delete, _scope, _noScope, _desc, _private, _template } from './get-args'
 import { askSingle } from 'reloquent'
 import getUsage from './usage'
 import signIn from '../lib/sign-in'
@@ -36,7 +36,9 @@ const getName = async (name) => {
 
     if (_check) return await runCheck(name)
 
-    const { token, scope: settingsScope, ...settings } = await signIn()
+    const {
+      token, scope: settingsScope,
+      template = 'mnpjs/package', ...settings } = await signIn()
 
     if (_delete) return await runDelete(token, settings.org, name)
 
@@ -45,7 +47,8 @@ const getName = async (name) => {
       ...settings,
     }, {
       name,
-      struct: _struct,
+      template: _template || template,
+      private: _private,
       token,
       description: _desc,
     })